Adds state to expand cards
