+++ src/components/Cards/styles.css
@@ -1,18 +1,9 @@
 h2 {
-    font-size: 1.5em;
+    font-size: 1.3rem;
 }
 
 h3 {
     font-size: 1em;
-}
-
-p {
-    white-space: nowrap;
-    overflow: hidden;
-}
-
-.card{
-    width: 30vh;
 }
 
 .card-img-top{
